Moved implementation into cpp file
